Changed relevant fields from int to long because videos with more than ${Int32.MaxValue} views or a file size larger than ${Int32.MaxValue} exist

+++ src/NYoutubeDL/Models/FormatDownloadInfo.cs
@@ -55,7 +55,7 @@ namespace NYoutubeDL.Models
 
         public string Ext { get; }
 
-        public int? Filesize { get; }
+        public long? Filesize { get; }
 
         public string Format { get; }
